Appendix Partial B 3.2 Exercises
Starting a new script meant that ‘Y’ was no longer set to a value, so
the script couldn’t find the sum and other information about Y based on
what it was set at earlier in Appendix B. I moved on to the next part
in 3.2 (“Algebra with Vectors”). The full B 3.2 is in the “Appendix B.3
Complete Exercises” commit.